refactor: refactored some logic out of the `PackageDb`
This clears up the code a bit and makes it easier to implement actual
Artifact support for STree and other `direct_url` dependencies
