+++ crates/rattler_installs_packages/src/index/direct_url/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod file;
+pub(crate) mod git;
+pub(crate) mod http;
