+++ crates/rattler_installs_packages/src/index/mod.rs
@@ -2,12 +2,13 @@
 
 mod file_store;
 
-mod git;
+mod direct_url;
+mod git_interop;
 pub mod html;
 mod http;
 mod package_database;
 
-pub use package_database::PackageDb;
+pub use package_database::{ArtifactRequest, PackageDb};
 
 pub use self::http::CacheMode;
 pub use html::parse_hash;
