+++ crates/rattler_installs_packages/src/types/artifact_name.rs
@@ -524,6 +524,13 @@ impl InnerAsArtifactName for SDistFilename {
     }
 }
 
+impl InnerAsArtifactName for STreeFilename {
+    fn try_as(name: &ArtifactName) -> Option<&Self> {
+        dbg!("get name");
+        name.as_stree()
+    }
+}
+
 #[cfg(test)]
 mod test {
     use super::*;
